Packageの整理+ViewFactoryの実装
