+++ Otap/OtapApp.swift
@@ -6,36 +6,16 @@
 //
 
 import SwiftUI
-import Combine
-import Components
-import Feature
-import FeatureImpl
+import ViewFactoryImpl
 import CoreImpl
-import UseCaseImpl
-import RepositoryImpl
+import Feature
 
 @main
 struct OtapApp: App {
     var body: some Scene {
         WindowGroup {
-//            VoiceQuizView(
-//                viewModel: VoiceQuizViewModelImpl<
-//                    ColorQuiz,
-//                    SoundEffectInteractor,
-//                    VoiceQuizInteractor<ColorQuiz, RepositoryImpl>
-//                >(
-//                    useCase: .init(
-//                        generator: ColorQuizCategory.level1.generator,
-//                        lastRecord: .init(
-//                            star1: false,
-//                            star2: false,
-//                            star3: false
-//                        )
-//                    )
-//                )
-//            )
             NavigationStack {
-                SelectCategoryView<SelectCategoryViewModel<ColorQuizCategory, SelectCategoryInteractor<RepositoryImpl>>>(viewModel: .init())
+                SelectCategoryView<SelectCategoryViewFactoryImpl<ColorQuizCategory>>(viewModel: .init())
             }
         }
     }
